Moves all basic stuff into a foundation package
