+++ components/src/jsMain/kotlin/dev/fritz2/components/appFrame.kt
@@ -2,6 +2,8 @@ package dev.fritz2.components
 
 import dev.fritz2.binding.SimpleHandler
 import dev.fritz2.binding.storeOf
+import dev.fritz2.components.foundations.Component
+import dev.fritz2.components.foundations.ComponentProperty
 import dev.fritz2.dom.html.Div
 import dev.fritz2.dom.html.RenderContext
 import dev.fritz2.dom.html.TextElement
