+++ components/src/jsMain/kotlin/dev/fritz2/components/checkboxes/componentSingle.kt
@@ -2,6 +2,7 @@ package dev.fritz2.components.checkboxes
 
 import dev.fritz2.binding.Store
 import dev.fritz2.components.*
+import dev.fritz2.components.foundations.*
 import dev.fritz2.dom.html.Input
 import dev.fritz2.dom.html.Label
 import dev.fritz2.dom.html.RenderContext
