+++ components/src/jsMain/kotlin/dev/fritz2/components/foundations/managed.kt
@@ -1,4 +1,4 @@
-package dev.fritz2.components
+package dev.fritz2.components.foundations
 
 import dev.fritz2.dom.HtmlTagMarker
 import dev.fritz2.dom.Tag
@@ -14,7 +14,51 @@ import kotlin.random.Random
 import kotlin.random.nextUInt
 
 /**
- * This interface marks
+ * This interface should be implemented by all components, that should appear at a specific, centralized node within
+ * the DOM tree. Think of [modal]s or [toast]s for example: They appear as overlays above all other screen elements,
+ * which is realized by a central parent node in the DOM.
+ *
+ * The render method has to be implemented in order to do the actual rendering of one component.
+ * This reduces the boilerplate code within the corresponding factory function(s):
+ * ```
+ * open class MyManagedComponent: ManagedComponent {
+ *      override fun render(...) {
+ *          // some content rendering
+ *      }
+ * }
+ *
+ * RenderContext.myComponent(
+ *      // most params omitted
+ *      build: MyManagedComponent.() -> Unit = {}
+ * ) {
+ *      MyComponent().apply(build).render(/* params */)
+ *      //                         ^^^^^^ ^^^^^^^^^^^^
+ *      //                         |      no reference to parent's `RenderContext` is needed!
+ *      //                         |      (compare this to `Component<T>` interface!
+ *      //                         |
+ *      //                         just start the rendering by one additional call!
+ * }
+ * ```
+ *
+ * To start the centralized rendering process, consider using the `init` block of a companion object:
+ * ```
+ * open class MyManagedComponent: ManagedComponent {
+ *     companion object {
+ *         init {
+ *             ManagedComponent.managedRenderContext("someId", Job()).apply {
+ *                 // do the central rendering (often surrounded by some central store, that manages the child
+ *                 // components added by the factory functions)
+ *             }
+ *         }
+ *     }
+ *
+ *      override fun render(...) {
+ *          // some content rendering
+ *      }
+ * }
+ * ```
+ *
+ * @see Component<T>
  */
 @HtmlTagMarker
 interface ManagedComponent<T> {
@@ -42,6 +86,7 @@ interface ManagedComponent<T> {
     companion object {
         /**
          * Gives a global [RenderContext] which is appended to the html body element.
+         *
          * @param id for [RenderContext] DOM element
          * @param job [Job] used in this [RenderContext]
          */
