+++ components/src/jsMain/kotlin/dev/fritz2/components/nav.kt
@@ -1,5 +1,8 @@
 package dev.fritz2.components
 
+import dev.fritz2.components.foundations.Component
+import dev.fritz2.components.foundations.ComponentProperty
+import dev.fritz2.components.foundations.DynamicComponentProperty
 import dev.fritz2.dom.Listener
 import dev.fritz2.dom.html.RenderContext
 import dev.fritz2.styling.*
