+++ components/src/jsMain/kotlin/dev/fritz2/components/navbar.kt
@@ -1,5 +1,7 @@
 package dev.fritz2.components
 
+import dev.fritz2.components.foundations.Component
+import dev.fritz2.components.foundations.ComponentProperty
 import dev.fritz2.dom.html.RenderContext
 import dev.fritz2.styling.StyleClass
 import dev.fritz2.styling.div
