+++ components/src/jsMain/kotlin/dev/fritz2/components/slider/component.kt
@@ -5,6 +5,7 @@ import dev.fritz2.binding.RootStore
 import dev.fritz2.binding.Store
 import dev.fritz2.binding.watch
 import dev.fritz2.components.*
+import dev.fritz2.components.foundations.*
 import dev.fritz2.dom.EventContext
 import dev.fritz2.dom.Window
 import dev.fritz2.dom.html.Div
