+++ components/src/jsMain/kotlin/dev/fritz2/components/textarea.kt
@@ -1,10 +1,10 @@
 package dev.fritz2.components
 
 import dev.fritz2.binding.Store
+import dev.fritz2.components.foundations.*
 import dev.fritz2.dom.html.RenderContext
 import dev.fritz2.dom.html.TextArea
 import dev.fritz2.dom.values
-import dev.fritz2.identification.uniqueId
 import dev.fritz2.styling.StyleClass
 import dev.fritz2.styling.name
 import dev.fritz2.styling.params.BasicParams
